Hide issue card PHI
