+++ src/applications/appeals/996/components/ShowIssuesList.jsx
@@ -11,11 +11,15 @@ export const ShowIssuesList = ({ issues }) => (
         </strong>
         <div>
           Decision date:{' '}
-          {getDate({
-            date:
-              issue.attributes?.approxDecisionDate || issue.decisionDate || '',
-            pattern: FORMAT_READABLE,
-          })}
+          <span className="dd-privacy-hidden">
+            {getDate({
+              date:
+                issue.attributes?.approxDecisionDate ||
+                issue.decisionDate ||
+                '',
+              pattern: FORMAT_READABLE,
+            })}
+          </span>
         </div>
       </li>
     ))}
